remove unnecessary casts when printing object addresses
casting pointers to 'long' causes warnings in 64-bit Windows builds (because 'long' is really 32-bit)

Note that this cast only appears in these three spots.
There are literally hundreds of places where we *don't* cast the pointer before passing it to ".x%lx"
Also, strnpointerid() is only used in two place (in s_print.c). We really need a general solution for the whole code base (which doesn't have to be a function!), so I went ahead and removed this function.

+++ src/g_rtext.c
@@ -434,7 +434,7 @@ static void rtext_senditup(t_rtext *x, int action, int *widthp, int *heightp,
             tcl/tk by escaping the close brace otherwise.  The GUI code
             drops the last character in the string. */
         sys_vgui("pdtk_text_new .x%lx.c {%s %s text} %d %d {%s } %d %s\n",
-            (long)canvas, x->x_tag, rtext_gettype(x)->s_name,
+            canvas, x->x_tag, rtext_gettype(x)->s_name,
             text_xpix(x->x_text, x->x_glist) + lmargin,
                 text_ypix(x->x_text, x->x_glist) + tmargin,
             escbuf,
